% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmx-package.r
\docType{package}
\name{dmx-package}
\alias{dmx-package}
\title{Distributed Matrix Extras}
\description{
TODO
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
