% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch_ddmatrix.r
\name{stitch_ddmatrix}
\alias{stitch_ddmatrix}
\title{stitch_ddmatrix}
\usage{
stitch_ddmatrix(x, by = "column", verify = TRUE)
}
\arguments{
\item{x}{Local data matrix.}

\item{by}{One of \code{"column"} or \code{"row"}. Are you stitching the data together
by columns or rows?}

\item{verify}{Should some checks be evaluated to (try to) make sure that the distributed
matrix makes sense?}
}
\value{
A ddmatrix.
}
\description{
Stitch together a distributed matrix from the local pieces. Only works for
1-d block distributions.
}
\details{
Assumptions:
\itemize{
  \item every process owns some of the data
  \itemize{
    \item the data should be distributable on BLACS grid 1 or 2
    \item some communications take place over \code{MPI_COMM_WORLD} (which
    corresponds to a \code{.pbd_env$SPMD.CT$comm} value of 0).
  }
  \item the data is laid out in 1-d block format
  \itemize{
    \item if \code{by == "column"} then all processes should have the same
    number of rows, and all but perhaps the last should have the same
    number of columns.
    \item if \code{by == "row"} then all processes should have the same
    number of columns, and all but perhaps the last should have the same
    number of rows.
  }
}
}
